refactor: move `BaseWorker` into application types
